requirements.openelm.training.lock.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cryptography==46.0.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
databricks-sdk==0.76.0
datasets==4.4.2
dill==0.4.0
docker==7.1.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huey==2.5.5
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.10.8
mlflow==3.8.1
mlflow-skinny==3.8.1
mlflow-tracing==3.8.1
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-triton==3.2.0+git4b3bb1f8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.1
tomli==2.3.0
torch==2.11.0.dev20260101+cu128
torchvision==0.25.0.dev20260101+cu128
tqdm==4.67.1
transformers==4.48.3
triton==3.6.0+git9844da95
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
